✨ feat(extensions): convert uv to directory layout with tests
